[step4] refactor(racingcar) test소스 package분리
