+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
